Fix datatype mismatch
Would result in out of order reading otherwise

+++ Binary-Data-Logger/UserDataType.h
@@ -3,6 +3,6 @@
 const int16_t ADC_DIM = 7;
 struct data_t {
   unsigned long time;
-  int8_t adc[ADC_DIM];
+  int16_t adc[ADC_DIM];
 };
 #endif  // UserDataType_h
